pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ieeT"
version = "0.9.0"
description = "arXiv 论文翻译工具 - 将英文 LaTeX 论文翻译成中文，保留数学公式和文档结构"
readme = "README.md"
requires-python = ">=3.10"
license = "GPL-3.0-or-later"
authors = [
  {name = "ieeT Team"},
]
keywords = ["arxiv", "latex", "translation", "chinese", "llm", "academic", "paper"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Text Processing :: Markup :: LaTeX",
  "Topic :: Scientific/Engineering",
]
dependencies = [
  "pylatexenc",
  "typer",
  "rich",
  "pyyaml",
  "pydantic",
  "requests",
  "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "ruff",
  "mypy",
  "build",
  "twine",
]

[project.urls]
Homepage = "https://github.com/ieeT/ieeT"
Repository = "https://github.com/ieeT/ieeT"

[project.scripts]
ieet = "ieet.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/ieet"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build]
exclude = [
  "/.github",
  "/docs",
  "/tests",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true
